CADC-14785 - initial renaming.
