+++ README.md
@@ -1,2 +1,2 @@
-# blank2caom2
-Starting point to build an application to generate CAOM2 Observations from FITS files.
+# apero2caom2
+An application to generate CAOM2 Observations from APERO-processed files.
