+++ apero2caom2/tests/conftest.py
@@ -2,7 +2,7 @@
 # ******************  CANADIAN ASTRONOMY DATA CENTRE  *******************
 # *************  CENTRE CANADIEN DE DONNÉES ASTRONOMIQUES  **************
 #
-#  (c) 2025.                            (c) 2025.
+#  (c) 2025.                             (c) 2025.
 #  Government of Canada                 Gouvernement du Canada
 #  National Research Council            Conseil national de recherches
 #  Ottawa, Canada, K1A 0R6              Ottawa, Canada, K1A 0R6
@@ -66,62 +66,37 @@
 # ***********************************************************************
 #
 
-"""
-This module implements the ObsBlueprint mapping, as well as the workflow 
-entry point that executes the workflow.
-"""
+from os.path import dirname, join, realpath
+from caom2pipe.manage_composable import Config, StorageName, TaskType
+import pytest
 
-from os.path import basename
-
-from caom2pipe import caom_composable as cc
-from caom2pipe import manage_composable as mc
+COLLECTION = 'APERO'
+SCHEME = 'cadc'
+PREVIEW_SCHEME = 'cadc'
 
 
-__all__ = [
-    'BlankMapping',
-    'BlankName',
-]
+@pytest.fixture()
+def test_config():
+    config = Config()
+    config.collection = COLLECTION
+    config.preview_scheme = PREVIEW_SCHEME
+    config.scheme = SCHEME
+    config.logging_level = 'INFO'
+    config.task_types = [TaskType.INGEST]
+    StorageName.collection = config.collection
+    StorageName.preview_scheme = config.preview_scheme
+    StorageName.scheme = config.scheme
+    StorageName.data_source_extensions = config.data_source_extensions
+    return config
 
 
-class BlankName(mc.StorageName):
-    """Naming rules:
-    - support mixed-case file name storage, and mixed-case obs id values
-    - support uncompressed files in storage
-    """
+@pytest.fixture()
+def test_data_dir():
+    this_dir = dirname(realpath(__file__))
+    fqn = join(this_dir, 'data')
+    return fqn
 
-    BLANK_NAME_PATTERN = '*'
-
-    def __init__(self, source_names):
-        super().__init__(source_names=source_names)
-
-    def is_valid(self):
-        return True
-
-
-class BlankMapping(cc.TelescopeMapping2):
-
-    def accumulate_blueprint(self, bp):
-        """Configure the telescope-specific ObsBlueprint at the CAOM model Observation level."""
-        self._logger.debug('Begin accumulate_bp.')
-        super().accumulate_blueprint(bp)
-
-        bp.set('Plane.calibrationLevel', 1)
-        bp.set('Plane.dataProductType', 'image')
-        bp.set('Artifact.productType', 'science')
-        bp.set('Artifact.releaseType', 'data')
-
-        bp.configure_position_axes((1, 2))
-        bp.configure_time_axis(3)
-        bp.configure_energy_axis(4)
-        bp.configure_polarization_axis(5)
-        bp.configure_observable_axis(6)
-        self._logger.debug('Done accumulate_bp.')
-
-    def update(self):
-        """Called to fill multiple CAOM model elements and/or attributes (an n:n relationship between TDM attributes 
-        and CAOM attributes).
-        """
-        return super().update()
-
-    def _update_artifact(self, artifact):
-        pass
+             
+@pytest.fixture()
+def change_test_dir(tmp_path, monkeypatch): 
+    monkeypatch.chdir(tmp_path)
