+++ scripts/apero_run_incremental.sh
